+++ Ilumination/Color.cpp
@@ -1,7 +1,7 @@
   #include "Color.h"
-  
+
   Color::Color(){}
-  Color::Color(float red, float green, float blue){
+  Color::Color(double red, double green, double blue){
     this->red = red;// & 0xFF000000;
     this->green = green;// & 0xFF000000;
     this->blue = blue;// & 0xFF000000;
