+++ Ilumination/Color.h
@@ -3,12 +3,12 @@
 #include "Reflect.h"
 #include "Intensity.h"
 struct Color {
-  float red;
-  float green;
-  float blue;
+  double red;
+  double green;
+  double blue;
   Color();
-  Color(float red, float green, float blue);
-  Color operator*(float right) {
+  Color(double red, double green, double blue);
+  Color operator*(double right) {
     return Color(red * right, green * right, blue * right);
   }
   Color operator*(Intensity right) {
